xinz4@illinois.edu/4

+++ src/main/java/ElectricPokemon.java
@@ -65,7 +65,7 @@ public class ElectricPokemon extends Pokemon {
         if (opponent.pokeType == PokemonType.ELECTRIC) {
             return false;
         }
-        if (specProb > (int) Math.random()) {
+        if (specProb > Math.random()) {
              System.out.print(this.specialtyAttack);
              opponent.setHitPoints(0);
              return true;
