+++ src/main/java/FirePokemon.java
@@ -64,7 +64,7 @@ public class FirePokemon extends Pokemon {
         if (opponent.pokeType == PokemonType.FIRE) {
             return false;
         }
-        if (specProb > (int) Math.random()) {
+        if (specProb > Math.random()) {
             System.out.print(this.specialtyAttack);
             opponent.setHitPoints(0);
             return true;
